convert all areas with Integer to use Long instead

+++ src/edu/sjsu/cmpe297/db/dao/CompanyDAO.java
@@ -52,7 +52,7 @@ public class CompanyDAO implements DatabaseAccessObject<Company>{
 		ResultSet rs = stmt.executeQuery();
 		
 		while (rs.next()) {
-			Integer facebookId = rs.getInt(1);
+			Long facebookId = rs.getLong(1);
 			String name = rs.getString(2);
 			Company c = new Company(facebookId, name);
 			companies.add(c);
@@ -67,12 +67,12 @@ public class CompanyDAO implements DatabaseAccessObject<Company>{
 	@Override
 	public Company get(Company data) throws SQLException {
 		PreparedStatement stmt = con.prepareStatement(READ_COMPANY_FOR_ID_CMD);
-		stmt.setInt(1, data.getFacebookId());
+		stmt.setLong(1, data.getFacebookId());
 		ResultSet rs = stmt.executeQuery();
 		
 		rs.next();
 		
-		Integer facebookId = rs.getInt(1);
+		Long facebookId = rs.getLong(1);
 		String name = rs.getString(2);
 		Company c = new Company(facebookId, name);
 		
@@ -85,7 +85,7 @@ public class CompanyDAO implements DatabaseAccessObject<Company>{
 	@Override
 	public void insert(Company data) throws SQLException {
 		PreparedStatement stmt = con.prepareStatement(INSERT_COMPANY_CMD);
-		stmt.setInt(1, data.getFacebookId());
+		stmt.setLong(1, data.getFacebookId());
 		stmt.setString(2, data.getName());
 		stmt.executeUpdate();
 		stmt.close(); 
@@ -94,9 +94,9 @@ public class CompanyDAO implements DatabaseAccessObject<Company>{
 	@Override
 	public void update(Company oldData, Company newData) throws SQLException {
 		PreparedStatement stmt = con.prepareStatement(UPDATE_COMPANY_CMD);
-		stmt.setInt(1, newData.getFacebookId());
+		stmt.setLong(1, newData.getFacebookId());
 		stmt.setString(2, newData.getName());
-		stmt.setInt(3, oldData.getFacebookId());
+		stmt.setLong(3, oldData.getFacebookId());
 		stmt.executeUpdate();
 		stmt.close();
 	}
@@ -104,7 +104,7 @@ public class CompanyDAO implements DatabaseAccessObject<Company>{
 	@Override
 	public void delete(Company data) throws SQLException {
 		PreparedStatement stmt = con.prepareStatement(DELETE_COMPANY_CMD);
-		stmt.setInt(1, data.getFacebookId());
+		stmt.setLong(1, data.getFacebookId());
 		stmt.executeUpdate();
 		stmt.close();
 	}
