+++ src/edu/sjsu/cmpe297/db/object/Views.java
@@ -3,38 +3,38 @@ package edu.sjsu.cmpe297.db.object;
 public class Views implements DatabaseObject {
 
 	// member variables
-	private Integer userId;
-	private Integer productId;
-	private Integer viewCount;
+	private Long userId;
+	private Long productId;
+	private Long viewCount;
 	
-	public Views(Integer userId, Integer productId, Integer viewCount) {
+	public Views(Long userId, Long productId, Long viewCount) {
 		this.userId = userId;
 		this.productId = productId;
 		this.viewCount = viewCount;
 	}
 	
 	// accessor methods
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 	
-	public Integer getProductId() {
+	public Long getProductId() {
 		return productId;
 	}
 	
-	public Integer getViewCount() {
+	public Long getViewCount() {
 		return viewCount;
 	}
 	
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 	
-	public void setProductId(Integer productId) {
+	public void setProductId(Long productId) {
 		this.productId = productId;
 	}
 	
-	public void setViewCount(Integer viewCount) {
+	public void setViewCount(Long viewCount) {
 		this.viewCount = viewCount;
 	}
